create private controller
